requirements.txt
uagents
uagents-core
requests
python-dotenv
fetchai
flask
